重构 Rotary Embedding 操作以保留输入张量的数据类型
- 更新了 `_apply_rotary_pos_emb_single` 和 `apply_rotary_pos_emb` 函数，确保输出张量与输入张量保持相同的数据类型。
- 修改了 `Gemma4TextRotaryEmbedding` 和 `Qwen3TextRotaryEmbedding` 类，去掉余弦和正弦张量的显式类型转换，提升张量处理的一致性。
- 新增测试用例，验证 Rotary Position Embedding 函数在多种配置下均能正确保留输入数据类型。

+++ src/anna/model/gemma4_text_model.py
@@ -67,7 +67,7 @@ def _apply_rotary_pos_emb_single(
     sin = sin.unsqueeze(1)
     rotary_dim = cos.shape[-1]
     hidden_rot, hidden_pass = hidden_states[..., :rotary_dim], hidden_states[..., rotary_dim:]
-    hidden_rot = (hidden_rot * cos) + (rotate_half(hidden_rot) * sin)
+    hidden_rot = ((hidden_rot * cos) + (rotate_half(hidden_rot) * sin)).to(dtype=hidden_states.dtype)
     return torch.cat((hidden_rot, hidden_pass), dim=-1)
 
 
@@ -513,7 +513,7 @@ class Gemma4TextRotaryEmbedding(nn.Module):
         emb = torch.cat((freqs, freqs), dim=-1)
         cos = emb.cos() * scaling
         sin = emb.sin() * scaling
-        return cos.to(dtype=hidden_states.dtype), sin.to(dtype=hidden_states.dtype)
+        return cos, sin
 
     def reset_runtime_buffers(self) -> None:
         for layer_type in self.layer_types:
